Fix -Wformat issue
Explicitly cast SensorType to its underlying type to avoid format
warning that only appears in some build configurations.

Fixes: 137234524
Test: ./run_tests.sh compiles cleanly and all tests pass

+++ core/sensor_request_manager.cc
@@ -103,7 +103,8 @@ bool SensorRequestManager::getSensorHandle(SensorType sensorType,
 
   bool sensorHandleIsValid = false;
   if (!isValidSensorType(sensorType)) {
-    LOGW("Querying for unknown sensor type %" PRIu8, sensorType);
+    LOGW("Querying for unknown sensor type %" PRIu8,
+         static_cast<uint8_t>(sensorType));
   } else {
     size_t sensorIndex = getSensorTypeArrayIndex(sensorType);
     sensorHandleIsValid = mSensorRequests[sensorIndex].isSensorSupported();
